# internal/spec/openapi.yaml
openapi: 3.1.0
info:
  title: hello
  version: 1.1.0
  description: Wake-on-LAN registry with live presence. Send magic packets, track
    devices, and subscribe to status changes over WebSocket.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /ws:
    get:
      tags: [System]
      summary: Live status stream
      description: >-
        WebSocket upgrade. The server pushes a StatusChange message every time
        the background monitor observes a reachability change. Clients should
        answer ping frames with pong; idle connections are closed after 60s
        without pong.
      operationId: liveStatus
      responses:
        '101':
          description: Switching protocols, status stream starts
        '426':
          description: Upgrade required, use a WebSocket client
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /wake:
    post:
      tags: [Wake]
      summary: Wake the default device
      description: Canonical endpoint. Sends a magic packet to DEFAULT_MAC.
      operationId: wakeDefault
      responses:
        '200':
          description: Magic packet sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WakeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Wake]
      summary: Wake the default device (legacy)
      deprecated: true
      description: Kept for backward compatibility. Prefer POST /wake because GET
        must not trigger side effects.
      operationId: wakeDefaultLegacy
      responses:
        '200':
          description: Magic packet sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WakeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /wake/{macAddress}:
    parameters:
      - $ref: '#/components/parameters/MacAddress'
    post:
      tags: [Wake]
      summary: Wake a device by MAC
      description: Canonical endpoint. Prefer over GET for the same reason as
        POST /wake.
      operationId: wakeMAC
      responses:
        '200':
          description: Magic packet sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WakeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Wake]
      summary: Wake a device by MAC (legacy)
      deprecated: true
      description: Kept for backward compatibility. Prefer POST.
      operationId: wakeMACLegacy
      responses:
        '200':
          description: Magic packet sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WakeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /devices:
    get:
      tags: [Devices]
      summary: List devices
      description: Filtered, sorted and paginated. The envelope always carries the total.
      operationId: listDevices
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DeviceStatus'
        - name: search
          in: query
          required: false
          description: Case-insensitive substring over id, name, MAC and IP
          schema:
            type: string
        - name: tag
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [name, status, last_seen, wake_count]
            default: name
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        '200':
          description: Current registry snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    post:
      tags: [Devices]
      summary: Register a device
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInput'
            examples:
              withPing:
                summary: Monitored host
                value:
                  id: pc-salon
                  name: PC Salon
                  mac: 00:11:22:33:44:55
                  ip: 192.168.1.50
                  ping_enabled: true
      responses:
        '201':
          description: Device registered, echo of the stored representation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      tags: [Devices]
      summary: Delete several devices
      description: Comma-separated ids. Unknown ids are reported, not fatal.
      operationId: bulkDeleteDevices
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
            example: pc-salon,pc-bureau
      responses:
        '200':
          description: Per-id report
          content:
            application/json:
              schema:
                type: object
                required: [deleted, not_found]
                properties:
                  deleted:
                    type: array
                    items:
                      type: string
                  not_found:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /devices/counts:
    get:
      tags: [Devices]
      summary: Registry totals by reachability
      operationId: deviceCounts
      responses:
        '200':
          description: Cheap polling totals
          content:
            application/json:
              schema:
                type: object
                required: [total, up, down, unknown]
                properties:
                  total:
                    type: integer
                  up:
                    type: integer
                  down:
                    type: integer
                  unknown:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /devices/export:
    get:
      tags: [Devices]
      summary: Download the registry
      description: Whole registry as a JSON array, served as attachment.
      operationId: exportDevices
      responses:
        '200':
          description: devices-export.json payload
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /devices/import:
    post:
      tags: [Devices]
      summary: Bulk-load devices
      description: >-
        Every entry is validated before anything is written, so a bad batch
        never leaves a partial registry. Modes are abort, skip and replace.
      operationId: importDevices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [devices]
              properties:
                devices:
                  type: array
                  items:
                    $ref: '#/components/schemas/DeviceInput'
                mode:
                  type: string
                  enum: [abort, skip, replace]
                  default: abort
      responses:
        '201':
          description: Created entries plus skipped ids
          content:
            application/json:
              schema:
                type: object
                required: [created, skipped]
                properties:
                  created:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
                  skipped:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /devices/wake-batch:
    post:
      tags: [Devices]
      summary: Wake several devices at once
      description: Select by explicit ids, by tag, or both; empty selection means the whole registry. Shares the wake cooldown. dry_run previews the target set without sending.
      operationId: wakeBatch
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
                tag:
                  type: string
                dry_run:
                  type: boolean
      responses:
        '200':
          description: Matched set plus per-device outcome
          content:
            application/json:
              schema:
                type: object
                required: [dry_run, matched, woken, failed]
                properties:
                  dry_run:
                    type: boolean
                  matched:
                    type: array
                    items:
                      type: string
                  woken:
                    type: array
                    items:
                      type: string
                  failed:
                    type: array
                    items:
                      type: object
                      required: [id, error]
                      properties:
                        id:
                          type: string
                        error:
                          type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /devices/{id}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    get:
      tags: [Devices]
      summary: Get a device
      operationId: getDevice
      responses:
        '200':
          description: Stored representation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Devices]
      summary: Replace a device
      description: Full replacement. The path id wins over any id in the body.
        Omitted fields reset to defaults (status defaults to unknown).
      operationId: updateDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInput'
      responses:
        '200':
          description: Stored representation after replacement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    patch:
      tags: [Devices]
      summary: Patch a device
      description: JSON merge patch, only the keys present are changed. The id is immutable here.
      operationId: patchDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:
                  type: string
                mac:
                  type: string
                  pattern: '^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$'
                ip:
                  type: string
                  format: ipv4
                status:
                  $ref: '#/components/schemas/DeviceStatus'
                ping_enabled:
                  type: boolean
                notes:
                  type: string
                  maxLength: 500
                tags:
                  type: array
                  maxItems: 10
                  items:
                    type: string
              additionalProperties: false
      responses:
        '200':
          description: Stored representation after patch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      tags: [Devices]
      summary: Delete a device
      operationId: deleteDevice
      responses:
        '204':
          description: Deleted, empty body
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /devices/{id}/clone:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    post:
      tags: [Devices]
      summary: Copy a device under a new id
      description: Counters and reachability reset; identity and network settings are kept.
      operationId: cloneDevice
      parameters:
        - name: new_id
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '201':
          description: The copy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /devices/{id}/wake:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    post:
      tags: [Devices]
      summary: Wake a registered device
      description: Looks the device up by id, then sends a magic packet to its MAC. With retries>0 the packet is resent until the device answers a ping or retries run out; retries x interval must stay within 8s (server write timeout) and the device needs an IP.
      operationId: wakeDevice
      parameters:
        - name: retries
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 5
            default: 0
        - name: interval
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
            default: 2
            description: Seconds between retry attempts
        - name: dry_run
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Validate only (values 1, true, yes); sends nothing, logs nothing
        - name: broadcast
          in: query
          required: false
          schema:
            type: string
            format: ipv4
          description: Override the computed directed broadcast for this wake
        - name: note
          in: query
          required: false
          schema:
            type: string
            maxLength: 140
          description: 'Free note stored on the history entry; a JSON body {"note": "..."} works too'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  maxLength: 140
                  description: Free note stored on the history entry
              additionalProperties: false
      responses:
        '200':
          description: Magic packet sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WakeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Retry needs a device IP, or a parameter is out of range
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          $ref: '#/components/responses/InternalError'
  /wake/oneshots:
    get:
      tags: [Devices]
      summary: List pending one-shot wakes
      description: Ephemeral delayed wakes, soonest first. Lost on restart.
      operationId: listOneshots
      responses:
        '200':
          description: Pending one-shots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OneshotList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Devices]
      summary: Schedule a one-shot wake
      description: Exactly one of device_id or mac, at most 24h ahead. Shares the wake throttle. Fires once as a manual wake with the optional note.
      operationId: createOneshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OneshotInput'
      responses:
        '201':
          description: One-shot armed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Oneshot'
        '400':
          description: Bad JSON, RFC 9457 body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Unknown device, bad time or note, RFC 9457 body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
  /wake/oneshots/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      tags: [Devices]
      summary: Cancel a pending one-shot
      operationId: deleteOneshot
      responses:
        '204':
          description: Cancelled, no body
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /history:
    get:
      tags: [Devices]
      summary: List recorded wakes
      description: Newest first. Filter by registry id, trigger, success, and time window; cap the page with limit.
      operationId: listHistory
      parameters:
        - name: device_id
          in: query
          required: false
          schema:
            type: string
        - name: trigger
          in: query
          required: false
          schema:
            type: string
            enum: [manual, schedule]
        - name: success
          in: query
          required: false
          schema:
            type: boolean
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Keep entries at or after this RFC 3339 time
        - name: before
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Keep entries strictly before this RFC 3339 time
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 50
      responses:
        '200':
          description: Wake log snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WakeHistory'
        '400':
          description: Bad limit, RFC 9457 body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Bad trigger, success, since or before value, RFC 9457 body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      tags: [Devices]
      summary: Purge old history entries
      description: Deletes every entry older than the required RFC 3339 before time. There is no undo.
      operationId: purgeHistory
      parameters:
        - name: before
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Purge count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurgeResult'
        '400':
          description: Missing before, RFC 9457 body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Bad before value, RFC 9457 body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /history/stats:
    get:
      tags: [Devices]
      summary: Wake statistics over retained history
      description: Totals, success rate, and per-trigger splits computed on the entries currently retained (bounded by HISTORY_CAP).
      operationId: historyStats
      responses:
        '200':
          description: Statistics snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /history/export:
    get:
      tags: [Devices]
      summary: Export history as CSV
      description: Downloads the retained entries as a history.csv attachment (id, at, device_id, mac, trigger, success, attempts, note, error).
      operationId: exportHistory
      responses:
        '200':
          description: CSV attachment
          content:
            text/csv:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /schedules/pause-all:
    post:
      tags: [Schedules]
      summary: Pause every schedule
      description: Disables all schedules at once and reloads the runner. Idempotent.
      operationId: pauseSchedules
      responses:
        '200':
          description: Pause report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PauseResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /schedules/resume-all:
    post:
      tags: [Schedules]
      summary: Resume every schedule
      description: Re-enables all schedules at once and reloads the runner. Idempotent.
      operationId: resumeSchedules
      responses:
        '200':
          description: Resume report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PauseResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
...
  schemas:
...
    DeviceStatus:
      type: string
      enum: [up, down, unknown]
...
    WakeResult:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: magic packet sent to AA:BB:CC:DD:EE:FF !
        attempts:
          type: integer
          description: Packets sent; only present on retry-until-up wakes
          example: 3
        up:
          type: boolean
          description: Whether the device answered a ping; only present on retry-until-up wakes
          example: true
      additionalProperties: false
...
    Problem:
      type: object
      description: RFC 9457 problem details. All errors share this envelope.
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: not found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: 'storage: device not found'
      additionalProperties: false
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token, RFC 9457 body
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API token from API_TOKEN. Browsers calling /ws use ?token= instead.